requirements.txt
pyinstaller==4.8
matplotlib==3.0.3
emoji==1.4.2
filetype==1.0.7
numpy==1.19.4
openpyxl==3.0.9
pandas==1.1.4
pathlib==1.0.1
phone-iso3166==0.3.7
Pillow==8.3.1
PySide2
pytz==2020.4
PyYAML==5.4.1
selenium==4.0.0
zipp==3.6.0
urllib3==1.26.2
requests
certifi==2019.11.28
pbkdf2==1.3
pycryptodome==3.11.0
